requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-MySQL==1.5.2
Flask-MySQLdb==1.0.1
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
frozenlist==1.3.3
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
multidict==6.0.4
mysql==0.0.3
mysql-connector-python==8.0.33
mysqlclient==2.1.1
openai==0.27.6
protobuf==3.20.3
PyMySQL==1.0.3
python-dateutil==2.8.2
requests==2.30.0
six==1.16.0
SQLAlchemy==2.0.14
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==2.0.2
Werkzeug==2.3.3
WTForms==3.0.1
yarl==1.9.2
